#256
- product-info: все интерфейсы перенесены в types (patch)
- product-info: ImageButton теперь принимает все пропсы StrokedSVG
