+++ src/common/components/product-info/parts/image-button.m.scss
@@ -1,7 +1,7 @@
 @use 'node_modules/@sima-land/ui-nucleons/colors';
 @use 'node_modules/@sima-land/ui-nucleons/breakpoints';
 
-.image-button {
+.root {
   width: 24px;
   height: 24px;
   display: block;
